GNL update

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: vimercie <vimercie@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 13:28:52 by vimercie          #+#    #+#             */
-/*   Updated: 2021/11/08 12:15:37 by vimercie         ###   ########lyon.fr   */
+/*   Updated: 2021/12/21 18:34:20 by vimercie         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
